move fetchSnapshot to containerStateManager
